Add Process object
Given a list of bids, which are the total number of candidates a department can accept, filter the number of roles down to that amount. If this number enables everyone to be matched, return them - otherwise, run another round. If two rounds aren't enough to match everyone, raise an Exception